Fix `Volume` binding sporadic (#373)
The Volume binding used to timeout since for the scheduler, first the
`Volume` and then the `VolumeClaim` was updated with the reference of
the other member.

This commit changes the `Volume.Status.Phase` from a plain `string` type
to a `VolumeCondition` of type `VolumeBound`. By doing so, the
controller can now check the last time the `VolumeBound` condition
transitioned and thus implement a grace period for binding (configurable
on the `Volume` controller).