Fix namespace deletion due to lacking protobuf
As mentioned in issue

https://github.com/kubernetes/kubernetes/issues/86666

if our types don't implement protobuf encoding and we still offer it as
part of our supported media types, namespaces using our types will hang
in deletion. This fixes it by providing a negotiated serializer that
supports all content types *but* protobuf. When either the issue is
fixed or we implement protobuf for our types the workaround can be
removed.