Machine state Terminated & power specification (#597)
This introduces a new state for a `Machine`, `Terminated`, that allows a
`MachinePool` implementor to signal if a `Machine` has been permanently
wiped so e.g. another controller on the top can act upon it.

Additionally, now a `Machine` can be requested to start / stop via
`Machine.Spec.Power`, supporting `On` and `Off` for now.