Implement network peering (#695)
This introduces support for network peering as described in
OEP-9.

The `Network` object is now extended with a `peerings` field that
contains the networks to peer with. If both networks reference each
other correctly, the peering status (reflected in `status.peerings`)
will be `Bound`.

Add test cases to verify peering. Update validation logic. Extend
external and internal types.