Fix network re-creation in machinebroker (#886)
Remove the network provider ID from the waiters map once an emit
successuflly finished. This solves the problem that in case the network
has been removed (e.g. via network GC controller), it is never
recreated. This causes a problem when brokering machine and NIC
resources as the OwnerRef can not be set on a non existing network
resource.