Update NIC status via machinepoollet
In the current implementation the NIC status {State,IPs} is not updated
by the `machinepoollet`. This leads to the problem that a NIC which is
`Available` is not being represented correctly in it's status. This PR
changes the `machinepoollet` logic in a way, that it populates effective
IPs to the NIC status and transitions the NIC state to `Available` if
the NIC could be provisioned successfully.